requirements.txt
# API 服务
fastapi>=0.109.0
uvicorn>=0.27.0
python-binance>=1.0.19
requests>=2.31.0
websockets>=12.0

# 数据处理
pandas>=2.2.0
pandas-ta>=0.3.14b0
numpy>=1.26.3

# 数据库
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# LLM
langchain>=0.1.0
openai>=1.10.0

# 任务调度
apscheduler>=3.10.4

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-settings>=2.1.0

# 日志
structlog>=24.1.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# 开发工具
black>=24.1.1
flake8>=7.0.0
isort>=5.13.2 